remove log

+++ interfaces/inner_api/app_manager/src/appmgr/ams_mgr_stub.cpp
@@ -103,7 +103,7 @@ void AmsMgrStub::CreateMemberFuncMap()
 
 int AmsMgrStub::OnRemoteRequest(uint32_t code, MessageParcel &data, MessageParcel &reply, MessageOption &option)
 {
-    if (code != IAmsMgr::Message::Get_BUNDLE_NAME_BY_PID) {
+    if (code != static_cast<uint32_t>(IAmsMgr::Message::Get_BUNDLE_NAME_BY_PID)) {
         HILOG_INFO("AmsMgrStub::OnReceived, code = %{public}u, flags= %{public}d.", code, option.GetFlags());
     }
     std::u16string descriptor = AmsMgrStub::GetDescriptor();
